+++ extras/tests/JsonVariant/isnull.cpp
@@ -70,19 +70,6 @@ TEST_CASE("JsonVariant::isNull()") {
     REQUIRE(variant.isNull() == true);
   }
 
-  SECTION("returns true for a shallow null copy") {
-    JsonDocument doc2(128);
-    variant.shallowCopy(doc2);
-    CHECK(variant.isNull() == true);
-  }
-
-  SECTION("returns false for a shallow array copy") {
-    JsonDocument doc2(128);
-    doc2[0] = 42;
-    variant.shallowCopy(doc2);
-    CHECK(variant.isNull() == false);
-  }
-
   SECTION("works with JsonVariantConst") {
     variant.set(42);